requirements.txt
requests
python-decouple
colorama